python-openai/requirements.txt
anyio==3.6.2
attrs==23.1.0
bcrypt==4.0.1
click==8.1.3
fastapi==0.95.1
h11==0.14.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
openapi==1.1.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.7
pyrsistent==0.19.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
